moved walmart transformer
